# DeviceArray.__iter__ returns DeviceArrays, without host sync