@hawkinsp 2021-09-30
I'm going to hold off merging this PR until after the ICLR deadline (Oct 5), since I think there's some chance of performance changes if nothing else.