@jakevdp 2021-12-11
A note: this change seems to have broken matplotlib's histogram function when a `DeviceArray` is passed to it (with most recent matplotlib and JAX releases):
```python
import matplotlib.pyplot as plt
import jax.numpy as jnp
import numpy as np

fix, ax = plt.subplots(1, 2, figsize=(10, 5))
ax[0].hist(np.arange(10))
ax[1].hist(jnp.arange(10));
```
![download-3](https://user-images.githubusercontent.com/781659/145657634-cabd9635-79b8-4244-a8f7-80e0c788cca6.png)

I believe this should be considered a matplotlib bug (see discussion in https://github.com/google/jax/issues/8899#issuecomment-991286825) but the new behavior here is what resulted in the problematic code path.