@jakevdp 2022-03-03
matplotlib issue fixed in https://github.com/matplotlib/matplotlib/issues/22016